+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - **aiken-lang**: Correctly infer Fuzzer & Sampler via type annotations when referring to foreign types. @KtorZ
 - **aiken-lang**: Allow type reification to pierce through Data aliases (e.g. Redeemer) holding lists, tuples or pairs, instead of crashing the compiler. @KtorZ
 - **aiken-lang**: Allow `Pair` to be used inline in Fuzzer's eDSL. @KtorZ
+- **aiken-lsp**: Fix module's inhabitant import quickfix for renamed modules. @KtorZ
+- **aiken-lsp**: Fix quickfix suggestion to use qualified for renamed modules. @KtorZ
 
 ## v1.1.16 - 2025-04-14